requirements.txt
Django==1.9.2
django-cors-headers==1.1.0
django-filter==0.13.0
django-suit==0.2.17
django-tinymce==2.3.0
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.10.2
paramiko==1.16.0
Pillow==3.1.1
psycopg2==2.6.1
pycrypto==2.6.1
